Added a new class EditorGrafico
-EditorGrafico class added in this  new commit.

This new class has an association relationship with the IGrafico interface, and its purpose is to act as a container for all created graphic objects.

Each Punto, Circulo, and Rectangulo object that is created is added to a list within this class. Currently, due to the limited nature of the exercise, EditorGrafico does not have an utility in the current state of the program. However, it could be used in a more complex program where it would be necessary to store new objects, as well as save or load the state of the EditorGrafico.