Single responsibility principle + improved try & catch
New version of the code. In this new version, the aim is to maintain the single responsibility principle. The Console.WriteLine have been removed from the classes, now they are only used in “main”. The Constructors now hold the exceptions so that the exceptions are handled as close as possible to where they occur.